requirements.txt
pdfplumber>=0.11.0
python-pptx>=1.0.0
